requirements.txt
streamlit
pandas
azure-storage-blob
azure-cosmos